backend/requirements.txt
acres==0.5.0
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cobble==0.1.4
configobj==5.0.9
configparser==7.2.0
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
looseversion==1.3.0
lxml==6.0.0
mammoth==1.9.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.30
pydot==4.0.1
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.4
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
traits==7.0.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
